init the contact us page
